Refactor: move nearby models to data.models
