+++ app/src/main/java/fr/free/nrw/commons/bookmarks/locations/BookmarkLocationsController.java
@@ -5,7 +5,7 @@ import java.util.List;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 
-import fr.free.nrw.commons.nearby.Place;
+import fr.free.nrw.commons.data.models.nearby.Place;
 
 @Singleton
 public class BookmarkLocationsController {
