+++ app/src/main/java/fr/free/nrw/commons/bookmarks/locations/BookmarkLocationsDao.java
@@ -19,8 +19,8 @@ import javax.inject.Provider;
 
 import fr.free.nrw.commons.location.LatLng;
 import fr.free.nrw.commons.nearby.Label;
-import fr.free.nrw.commons.nearby.Place;
-import fr.free.nrw.commons.nearby.Sitelinks;
+import fr.free.nrw.commons.data.models.nearby.Place;
+import fr.free.nrw.commons.data.models.nearby.Sitelinks;
 import timber.log.Timber;
 
 import static fr.free.nrw.commons.bookmarks.locations.BookmarkLocationsContentProvider.BASE_URI;
