+++ app/src/main/java/fr/free/nrw/commons/data/models/nearby/Place.java
@@ -1,17 +1,16 @@
-package fr.free.nrw.commons.nearby;
+package fr.free.nrw.commons.data.models.nearby;
 
 import android.net.Uri;
 import android.os.Parcel;
 import android.os.Parcelable;
 
-import android.text.TextUtils;
 import androidx.annotation.Nullable;
 
-import fr.free.nrw.commons.nearby.NearbyController.NearbyPlacesInfo;
+import fr.free.nrw.commons.nearby.Label;
 import org.apache.commons.lang3.StringUtils;
 
 import fr.free.nrw.commons.location.LatLng;
-import fr.free.nrw.commons.nearby.model.NearbyResultItem;
+import fr.free.nrw.commons.data.models.nearby.model.NearbyResultItem;
 import fr.free.nrw.commons.utils.PlaceUtils;
 import timber.log.Timber;
