+++ app/src/main/java/fr/free/nrw/commons/data/models/nearby/Sitelinks.java
@@ -1,4 +1,4 @@
-package fr.free.nrw.commons.nearby;
+package fr.free.nrw.commons.data.models.nearby;
 
 import android.net.Uri;
 import android.os.Parcel;
