+++ app/src/main/java/fr/free/nrw/commons/data/models/nearby/model/NearbyResponse.kt
@@ -1,0 +1,3 @@
+package fr.free.nrw.commons.data.models.nearby.model
+
+class NearbyResponse(val results: NearbyResults)
