+++ app/src/main/java/fr/free/nrw/commons/data/models/nearby/model/NearbyResults.kt
@@ -1,0 +1,5 @@
+package fr.free.nrw.commons.data.models.nearby.model
+
+import fr.free.nrw.commons.data.models.nearby.model.NearbyResultItem
+
+class NearbyResults(val bindings: List<NearbyResultItem>)
