+++ app/src/main/java/fr/free/nrw/commons/data/models/nearby/model/ResultTuple.kt
@@ -1,4 +1,4 @@
-package fr.free.nrw.commons.nearby.model
+package fr.free.nrw.commons.data.models.nearby.model
 
 import com.google.gson.annotations.SerializedName
