+++ app/src/main/java/fr/free/nrw/commons/nearby/NearbyBaseMarker.java
@@ -8,6 +8,7 @@ import com.mapbox.mapboxsdk.annotations.BaseMarkerOptions;
 import com.mapbox.mapboxsdk.annotations.Icon;
 import com.mapbox.mapboxsdk.annotations.IconFactory;
 import com.mapbox.mapboxsdk.geometry.LatLng;
+import fr.free.nrw.commons.data.models.nearby.Place;
 import java.util.Objects;
 
 public class NearbyBaseMarker extends BaseMarkerOptions<NearbyMarker, NearbyBaseMarker> {
