+++ app/src/main/java/fr/free/nrw/commons/nearby/NearbyController.java
@@ -12,6 +12,7 @@ import androidx.vectordrawable.graphics.drawable.VectorDrawableCompat;
 import com.mapbox.mapboxsdk.annotations.IconFactory;
 import com.mapbox.mapboxsdk.annotations.Marker;
 import fr.free.nrw.commons.R;
+import fr.free.nrw.commons.data.models.nearby.Place;
 import fr.free.nrw.commons.location.LatLng;
 import fr.free.nrw.commons.utils.UiUtils;
 import java.util.ArrayList;
