--- app/src/main/java/fr/free/nrw/commons/nearby/model/NearbyResponse.kt
@@ -1,3 +1,0 @@
-package fr.free.nrw.commons.nearby.model
-
-class NearbyResponse(val results: NearbyResults)
