--- app/src/main/java/fr/free/nrw/commons/nearby/model/NearbyResults.kt
@@ -1,3 +1,0 @@
-package fr.free.nrw.commons.nearby.model
-
-class NearbyResults(val bindings: List<NearbyResultItem>)
