+++ app/src/main/java/fr/free/nrw/commons/upload/ImageProcessingService.java
@@ -6,7 +6,7 @@ import static fr.free.nrw.commons.utils.ImageUtils.IMAGE_OK;
 
 import android.content.Context;
 import fr.free.nrw.commons.media.MediaClient;
-import fr.free.nrw.commons.nearby.Place;
+import fr.free.nrw.commons.data.models.nearby.Place;
 import fr.free.nrw.commons.utils.ImageUtils;
 import fr.free.nrw.commons.utils.ImageUtilsWrapper;
 import io.reactivex.Single;
