+++ app/src/test/kotlin/fr/free/nrw/commons/ModelFunctions.kt
@@ -4,8 +4,8 @@ import fr.free.nrw.commons.Media
 import fr.free.nrw.commons.category.CategoryItem
 import fr.free.nrw.commons.location.LatLng
 import fr.free.nrw.commons.nearby.Label
-import fr.free.nrw.commons.nearby.Place
-import fr.free.nrw.commons.nearby.Sitelinks
+import fr.free.nrw.commons.data.models.nearby.Place
+import fr.free.nrw.commons.data.models.nearby.Sitelinks
 import fr.free.nrw.commons.upload.structure.depictions.DepictedItem
 import fr.free.nrw.commons.wikidata.model.DepictSearchItem
 import org.wikipedia.wikidata.*
@@ -86,7 +86,17 @@ fun place(
     pic: String = "pic",
     exists: Boolean = false
 ): Place {
-    return Place(lang, name, label, longDescription, latLng, category, siteLinks, pic, exists)
+    return Place(
+        lang,
+        name,
+        label,
+        longDescription,
+        latLng,
+        category,
+        siteLinks,
+        pic,
+        exists
+    )
 }
 
 fun entityId(wikiBaseEntityValue: WikiBaseEntityValue = wikiBaseEntityValue()) =
