+++ app/src/test/kotlin/fr/free/nrw/commons/nearby/NearbyControllerTest.kt
@@ -5,6 +5,8 @@ import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.anyOrNull
 import com.nhaarman.mockitokotlin2.eq
 import fr.free.nrw.commons.TestCommonsApplication
+import fr.free.nrw.commons.data.models.nearby.Place
+import fr.free.nrw.commons.data.models.nearby.Sitelinks
 import fr.free.nrw.commons.location.LatLng
 import fr.free.nrw.commons.nearby.NearbyController.loadAttractionsFromLocationToBaseMarkerOptions
 import fr.free.nrw.commons.nearby.NearbyController.updateMarkerLabelListBookmark
